Import scenarios via upload of .json file
The administrator can now import scenarios by uploading the .json
file, that was created using the export function from earlier commits.

Importing an old version of a scenario will override changes but won't
delete elements that were added in the meantime. The imported version
of the scenario will be merged into the existing one.