Integrating reveal.js to create slideshow overviews
Added css and js files to integrate reveal.js.
A layout file was created that includes those files and can be used
instead of the normal layout.html file when creating presentation views.

A view was added that uses reveal.js to show the process of each user in
a scenario. The view (views/overview/reveal_progress.html) can be used as an example on how to use reveal.js
in the monitutor environment.